refactor : 정리
